+++ common/src/lib.rs
@@ -14,5 +14,5 @@ mod service_method;
 
 pub use crate::connection::{ConnectionTrait, ReadonlyConnection};
 pub use crate::message::{DecodableMessage, EncodableMessage, ReceivableMessage, SendableMessage};
-pub use crate::service_method::ServiceMethodRequest;
+pub use crate::service_method::{ServiceMethodRequest, ServiceNotification};
 pub use net::{JobId, NetMessageHeader, RawSteamId};