+++ src/message.rs
@@ -15,7 +15,7 @@ use std::io::{Cursor, Read, Write};
 use std::{fmt::Debug, io::ErrorKind};
 use steam_vent_common::{
     DecodableMessage, EncodableMessage, NetMessageHeader, ReceivableMessage, SendableMessage,
-    ServiceMethodRequest,
+    ServiceMethodRequest, ServiceNotification,
 };
 use steam_vent_proto_common::{MsgKind, RpcMessage};
 use steam_vent_proto_steam::enums_clientserver::EMsg;
@@ -215,14 +215,14 @@ pub(crate) struct ServiceMethodMessage<Request: Debug>(pub Request);
 
 impl<Request: ServiceMethodRequest + Debug> EncodableMessage for ServiceMethodMessage<Request> {
     fn write_body<W: Write>(&self, mut writer: W) -> Result<(), IoError> {
-        trace!("writing body of protobuf message {:?}", Self::KIND);
+        trace!("writing body of protobuf message {:?}", self.kind());
         self.0
             .write(&mut writer)
             .map_err(|_| IoError::from(ErrorKind::InvalidData))
     }
 
     fn encode_size(&self) -> usize {
-        self.0.compute_size() as usize
+        self.0.encode_size()
     }
 
     fn process_header(&self, header: &mut NetMessageHeader) {
@@ -230,28 +230,15 @@ impl<Request: ServiceMethodRequest + Debug> EncodableMessage for ServiceMethodMe
     }
 }
 
-impl<Request: ServiceMethodRequest + Debug> DecodableMessage for ServiceMethodMessage<Request> {
-    fn read_body(data: BytesMut, _header: &NetMessageHeader) -> Result<Self, IoError> {
-        trace!("reading body of protobuf message {:?}", Self::KIND);
-        Request::parse(&mut data.reader()).map(ServiceMethodMessage)
-    }
-}
-
-impl<Request: ServiceMethodRequest + Debug> ReceivableMessage for ServiceMethodMessage<Request> {
-    type KindEnum = EMsg;
-    const KIND: Self::KindEnum = EMsg::k_EMsgServiceMethodCallFromClient;
-    const IS_PROTOBUF: bool = true;
-}
-
 impl<Request: ServiceMethodRequest + Debug> SendableMessage for ServiceMethodMessage<Request> {
     type KindEnum = EMsg;
 
     fn kind(&self) -> Self::KindEnum {
-        Self::KIND
+        EMsg::k_EMsgServiceMethodCallFromClient
     }
 
     fn is_protobuf(&self) -> bool {
-        Self::IS_PROTOBUF
+        true
     }
 }
 
@@ -304,13 +291,13 @@ pub(crate) struct ServiceMethodNotification {
 }
 
 impl ServiceMethodNotification {
-    pub fn into_notification<Request: ServiceMethodRequest>(self) -> Result<Request, NetworkError> {
-        if self.job_name == Request::REQ_NAME {
+    pub fn into_notification<Request: ServiceNotification>(self) -> Result<Request, NetworkError> {
+        if self.job_name == Request::NOTIFICATION_NAME {
             Ok(Request::parse(&mut self.body.reader())
                 .map_err(|e| MalformedBody::new(Self::KIND, e))?)
         } else {
             Err(NetworkError::DifferentServiceMethod(
-                Request::REQ_NAME,
+                Request::NOTIFICATION_NAME,
                 self.job_name,
             ))
         }